Move settings dialog into separate file
